requirements.txt
Django>=4.2.0
pymysql>=1.0.2
psycopg2-binary>=2.9.6
pymongo>=4.3.3
pyodbc>=4.0.39
sqlalchemy>=2.0.0
django-crispy-forms>=2.0
python-dotenv>=1.0.0